Projeto_MercadoLivre_v2/python/requirements.txt
absl-py
amply
astroid
black
certifi
charset-normalizer
click
contourpy
coverage
cupy-cuda12x
cycler
dill
docutils
fastrlock
flake8
fonttools
idna
immutabledict
iniconfig
isort
keyboard
kiwisolver
llvmlite
matplotlib
mccabe
mypy_extensions
numba
numpy
ortools
packaging
pandas
pathspec
pillow
platformdirs
pluggy
prettytable
protobuf
PuLP
pycodestyle
pyflakes
pylint
pyparsing
pytest
pytest-cov
python-dateutil
python-dotenv
pytz
readchar
requests
scipy
setuptools
six
termcolor
tomlkit
tqdm
tzdata
urllib3
wcwidth
wheel
watchdog